requirements.txt
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
groq==0.9.0
chromadb==0.5.0
streamlit==1.40.0
pypdf==5.0.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
huggingface-hub==0.23.0
sentence-transformers==3.0.0
wikipedia==1.4.0
duckduckgo-search==3.9.13
pyttsx3==2.90
SpeechRecognition==3.10.0
gtts==2.4.0